04-belge-yonetim/elektronik-belge-yonetim-sistemi/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
python-magic==0.4.27
Pillow==10.1.0
django-filter==23.3
django-storages==1.14.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
django-taggit==5.0.1
django-mptt==0.15.0
boto3==1.29.7
cryptography==41.0.7
reportlab==4.0.7
openpyxl==3.1.2